refactor bloom_math
this doesn't have to be part of the main for the code and they contain quite a bit of complexity and interdependency so I factored them out.
