+++ src/bloom_filter/bloom_math.rs
@@ -1,0 +1,45 @@
+use super::BloomFilter;
+
+impl BloomFilter {
+    // n: number of items in filter. p: false positive rate
+    // m: number of bits in filter. k: number of hashers
+    // n = ceil(m / (-k / log(1 - exp(log(p) / k))))
+    // p = pow(1 - exp(-k / (m / n)), k)
+    // m = ceil((n * log(p)) / log(1 / pow(2, log(2))));
+    // k = round((m / n) * log(2));
+
+    pub fn optimal_number_of_hashers(size_in_bytes: usize, expected_elements: usize) -> usize {
+        let expected_elements = expected_elements as f64;
+        let size_in_bits = (size_in_bytes * 8) as f64;
+        let k = (size_in_bits / expected_elements) * (2.0f64.ln());
+        k.ceil() as usize
+    }
+
+    pub fn prob_of_false_positive(
+        size_in_bytes: usize,
+        expected_elements: usize,
+        num_hashers: usize,
+    ) -> f64 {
+        let k = num_hashers as f64;
+        let m = (size_in_bytes * 8) as f64;
+        let n = expected_elements as f64;
+        (1.0 - (1.0 - (1.0 / m)).powf(k * n)).powf(k)
+    }
+
+    pub fn suggest_size_in_bytes(
+        expected_elements: usize,
+        desired_false_positive_rate: f64,
+    ) -> usize {
+        let mut size_in_bytes = 1024 * 1024;
+        while size_in_bytes < usize::MAX / 2
+            && Self::prob_of_false_positive(
+                size_in_bytes,
+                expected_elements,
+                Self::optimal_number_of_hashers(size_in_bytes, expected_elements),
+            ) > desired_false_positive_rate
+        {
+            size_in_bytes *= 2;
+        }
+        size_in_bytes
+    }
+}
